Modified Enrollments.groovy

+++ grails-app/domain/gradadvising/Enrollments.groovy
@@ -23,8 +23,8 @@ class Enrollments {
 		double totalLoad = 0;
 		grade.each()
 		{
-			result = it.subject.unitLoad * it.getEquivalentGrade()
-			totalLoad += it.subject.unitLoad
+			result = (double)it.subject.unitLoad * it.getEquivalentGrade()
+			totalLoad += (double)it.subject.unitLoad
 		}
 		return result/totalLoad;
 	}
